requirements.txt
rembg[cpu]
Pillow
numpy
scipy
